fix(tests): silence MSVC C4267 size_t narrowings (#23 Phase 2b) (#26)
Eight call sites across three test files where size_t flowed into a
narrower destination, triggering MSVC C4267. Two patterns:

(1) test_memory_controller.cpp:140,147 and test_dma_engine_process.cpp:365,370
    store count_events()'s size_t return into a local 'int'. Widened the
    locals to size_t so the type matches the source. No casting needed,
    and the subsequent REQUIRE(x == 1) still compiles cleanly.

(2) test_dma_memory_noc_integration.cpp:460,498,537,551 compute a node
    id from a size_t loop variable and assign it to DMATransfer::dst_id,
    which is uint32_t (dma_engine_interface.hpp:60). The truncation is
    safe in practice (these tests use 4x4 = 16-node grids; the value
    never approaches 2^32) but should be made explicit with
    static_cast<uint32_t> at the assignment.

Verified locally on Linux gcc release:
  test_memory_controller:               12/12 cases,  51/51 assertions
  test_dma_engine_process:              13/13 cases,  54/54 assertions
  dma_memory_noc_integration_test:       8/8 cases,  61/61 assertions

Refs #23

+++ tests/dma/test_dma_memory_noc_integration.cpp
@@ -457,7 +457,7 @@ TEST_CASE("DMA System with placement", "[dma][system][integration]") {
             transfer.src_type = DMAMemoryType::DEVICE_MEMORY;
             transfer.src_addr = 0x10000 * row;
             transfer.dst_type = DMAMemoryType::L3_TILE;
-            transfer.dst_id = row * 4 + 3;  // East edge of each row
+            transfer.dst_id = static_cast<uint32_t>(row * 4 + 3);  // East edge of each row
             transfer.size = 64;
 
             system.west_dmas[row]->submit(transfer, [&completed_count]() {
@@ -495,7 +495,7 @@ TEST_CASE("DMA System with placement", "[dma][system][integration]") {
             transfer.src_type = DMAMemoryType::DEVICE_MEMORY;
             transfer.src_addr = 0x20000 * col;
             transfer.dst_type = DMAMemoryType::L3_TILE;
-            transfer.dst_id = 12 + col;  // South edge
+            transfer.dst_id = static_cast<uint32_t>(12 + col);  // South edge
             transfer.size = 64;
 
             system.north_dmas[col]->submit(transfer, [&completed_count]() {
@@ -534,7 +534,7 @@ TEST_CASE("DMA System with placement", "[dma][system][integration]") {
             transfer.src_type = DMAMemoryType::DEVICE_MEMORY;
             transfer.src_addr = 0x10000 * row;
             transfer.dst_type = DMAMemoryType::L3_TILE;
-            transfer.dst_id = row * 4 + 2;
+            transfer.dst_id = static_cast<uint32_t>(row * 4 + 2);
             transfer.size = 64;
 
             system.west_dmas[row]->submit(transfer, [&west_completed]() {
@@ -548,7 +548,7 @@ TEST_CASE("DMA System with placement", "[dma][system][integration]") {
             transfer.src_type = DMAMemoryType::DEVICE_MEMORY;
             transfer.src_addr = 0x20000 * col;
             transfer.dst_type = DMAMemoryType::L3_TILE;
-            transfer.dst_id = 8 + col;
+            transfer.dst_id = static_cast<uint32_t>(8 + col);
             transfer.size = 64;
 
             system.north_dmas[col]->submit(transfer, [&north_completed]() {
